requirements.txt
streamlit==1.26.0
requests==2.31.0
google-genai==0.3.0